+++ src/App.scss
@@ -52,6 +52,10 @@ body {
     padding: 5px 15px;
 }
 
+.hidden {
+    display: none;
+}
+
 @media screen and (min-width: 620px) {
     body {
       font-size: 1.6rem;
