--- src/components/Description/Description.module.scss
@@ -1,33 +1,0 @@
-@use "/src/styles/variables";
-
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    min-height: calc(100vh - #{variables.$header});
-    max-width: 500px;
-    margin: 0 auto;
-    padding: 20px;
-}
-
-.header_text {
-    margin: 25px 0;
-    font-size: 3.2rem;
-    font-family: 'Playfair Display', serif;
-    font-weight: 600;
-    text-transform: uppercase;
-}
-
-.description_text {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 25px;
-}
-
-.text {
-    margin-bottom: 10px;
-    text-align: center;
-}
