+++ src/components/Home/Home.module.scss
@@ -1,10 +1,33 @@
+@use "/src/styles/variables";
+
 .container {
-    width: 100%;
-    max-width: 1280px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    min-height: calc(100vh - #{variables.$header});
+    max-width: 500px;
     margin: 0 auto;
-    padding: 20px 0 40px;
+    padding: 20px;
 }
 
-.message {
-    padding: 20px 0;
+.header_text {
+    margin: 25px 0;
+    font-size: 3.2rem;
+    font-family: 'Playfair Display', serif;
+    font-weight: 600;
+    text-transform: uppercase;
+}
+
+.description_text {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 25px;
+}
+
+.text {
+    margin-bottom: 10px;
+    text-align: center;
 }
