+++ src/components/SearchPage/SearchPage.module.scss
@@ -1,0 +1,10 @@
+.container {
+    width: 100%;
+    max-width: 1280px;
+    margin: 0 auto;
+    padding: 20px 0 40px;
+}
+
+.message {
+    padding: 20px 0;
+}
